+++ java_hometasks/IPlate.java
@@ -1,0 +1,6 @@
+package java_hometasks;
+
+public interface IPlate {
+
+	
+}
