+++ java_hometasks/Plate.java
@@ -1,0 +1,5 @@
+package java_hometasks;
+
+public class Plate implements IPlate {
+
+}
